rename utils/sriUtils to util/sri
