+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 export { verifyX5C } from './verifyX5C';
 export { getSdJwtVcMetadata } from './getSdJwtVcMetadata';
 export { resolveWellKnownUrl } from './resolveWellKnownUrl';
-export * from './sriUtils';
+export * from './sri';
 export * from './util';
